Add checks to force user to be logged in

+++ src/main/java/it/engineering/web/test/action/producer/DeleteConfirmAction.java
@@ -4,11 +4,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import it.engineering.web.test.action.AbstractAction;
+import it.engineering.web.test.constants.Constants;
 
 public class DeleteConfirmAction extends AbstractAction {
 
 	@Override
 	public String executeRequest(HttpServletRequest request, HttpServletResponse response) {	
+		if (request.getSession(true).getAttribute("user")==null) {
+			return Constants.PAGE_LOGIN;
+		}
 		return getProducerService().deleteConfirmPressed(request);
 	}
